model-runner/python/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
boto3>=1.28.0
requests>=2.31.0
tqdm>=4.65.0
datasets>=2.14.0
av>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
